--- ui/src/components/ComponentInjector/Clipboard/utils/test.ts
@@ -1,8 +1,0 @@
-interface ClipboardProps {
-	text: string;
-	clipboardComponent: React.ElementType;
-	className?: string;
-	label?: string;
-}
-
-export type { ClipboardProps };
