+++ ui/src/pages/Converter/Svg/helper.ts
