+++ ui/src/pages/Generator/Data/constants.ts
