+++ ui/src/pages/Generator/Data/utils.ts
