+++ ui/src/pages/Generator/Image/constants.ts
