+++ ui/src/pages/Generator/Image/helper.ts
